+++ armonaut/auth/models.py
@@ -12,16 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from armonaut.db import Model
 from sqlalchemy import Column, String, DateTime, Boolean
+from sqlalchemy import sql
+from armonaut.db import Model
 
 
 class User(Model):
     __tablename__ = 'users'
 
     email = Column(String(96), nullable=False, index=True, unique=True)
-    password = Column(String(60), nullable=False)
+    password = Column(String(128), nullable=False)
 
-    email_verified = Column(Boolean, default=False, nullable=False)
-    email_verify_code = Column(String(32), default=None)
-    email_verify_expires = Column(DateTime, default=None)
+    is_staff = Column(Boolean, nullable=False)
+    is_superuser = Column(Boolean, nullable=False)
+
+    joined_date = Column(DateTime, nullable=False, server_default=sql.func.now())
+    password_date = Column(DateTime, nullable=False, server_default=sql.func.now())
+    last_login = Column(DateTime, nullable=False, server_default=sql.func.now())